requirements.txt
iniconfig==2.0.0
numpy==1.26.1
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
warp-lang==0.10.1
